Add ScuttleButt UDP server
This patch adds a new `server` module, providing functionality for two
ScuttleButt instances on different machines to stay in sync.

The new `ScuttleServer` struct provides a UDP server which runs in the
background once spawned with `ScuttleServer::spawn`, automatically
replying to incoming gossip using the shared `ScuttleButt` mutex.

To propagate gossip to other servers, the `ScuttleServer::gossip` method
allows transmitting the current state to other instances. While it would
be possible to keep track of all incoming and outgoing conversations, to
allow for broadcasting updates to all known nodes, this is currently
still left to the consumer.

Fixes #2.