remove invalid ignore_oversized_payload test
Introduced in #4, this test has been outdated since the reception
buffer sized has been increased to the full UDP MTU.

As is, it sends an empty Syn message before sending another valid
Syn message. The assert checks for an SynAck message, that is the
response to the first packet instead of the second.

Now that peers check the cluster name property of Syn messages,
the first packet gets a BadCluster response, which fails the assert.