- removed the use of string everywhere for addresses
The string were hiding several bugs.
This PR partially addresses it.

Before this PR, all addresses were `String`, that were
eventually resolved via `ToSocketAddrs`.

There are several cons and one bug associated with this approach.
Readability was hurt and API contract enforcing was a bit broken.
For instance, a user could have supplied an unresolved hostname as a gossip address.
As a result, all calls to that node would have triggered a DNS request.

In addition, the logic that checks for peers to gossip with takes contains
some logic that relies on equality. For instance, we check if no seed was
already picked, before artificially adding a seed candidate.

That logic is broken, because the seed addrs are likely to be unresolved names
while peer address are likely to be socket addresses.

We switched all address to socket address except seeds.
Supplying a hostname as a socket addr is something that is very useful for k8s
for instance.

Internally, we just do slow polling over the name resolution of this seed hostname
to detect eventual new seed servers.

- added node_id, public_addr option in chitchat test program

- simplified subprocess kill logic in unit test

- Seeding the random generator with the thread local generator (itself has been seed using system entropy).
This might help speed up the unit test a little bit.