Adding a transport abstraction.
The point of this transport layer is to:
- make it easy to simulate what happens if X% of the message are dropped, or if they are delayed.
- avoid having to juggle with UDP port allocation in unit test, and safely simulate 100 of nodes.
measure metrics such as the amount of data exchanged during simulations.
- slightly simplify the code of the server.

This PR also:
- Stops updating heartbeat on all ack messages.
While this was not hurting correctness, in the same round of gossip, we would send 3 different gossip to peers A,B,C.
This would generate meaningless update messages within the cluster.
- Adds a lot of performance tests simulating chitchat and checking that it is behaving correctly.

Closes #53
Closes #54
Closes #55