Change in chitchat gossiping priority.
Following the original paper, chitchat currently first shares nodes with the highest number of stale values.

As a side effect, nodes that are not emitting many KVs are gossiped last.
In quickwit, under a little bit of load (1000 indexes on 10 indexer), it has a very dire effect.

Indexer that reconnect have to gossip the entire cluster state (~10MB) before being able to get any
information about the metastore.
Knowing at least one node with the metastore service is required for nodes to declare themselves as live.

This PR changes the gossip order.
It prioritizes nodes for which the node that originated the Syn message
has not receives any KV yet.

This is identified by the fact that either the node was not part of the
digest at all, or the floor_version is equal to 0.

The latter (floor_version == 0) should not happen today, but this case
is done in preparation for another PR updating heartbeat on Syn.