Use the heartbeat in digest for cluster membership. (#123)
Right now we send the heartbeat of all of the known nodes in the cluster
during digests.

We can use that information for cluster membership.
This will allow nodes to detect new member, join a cluster, and detect
failure considerably faster.

This PR also removes heartbeat from delta message.
Instead it joins the last_gc_version, to fix a bug introduced in #122.

Bugfix:
Avoid updating the time of death of a node when it is redetected as
faulty.

We remove outselves from the list of seeds. Before nodes were regularly gossipping with themselves.

Code improvement:
In the failure detector when we have few samples. We now use additive
smoothing.