Resetted nodes were briefly identified as alive.
The reason is that we call report_heartbeat in the reset procedure.
The purpose of this call is to make sure that we have a record in the failure detector for the node.

This PR just creates the failure detector entry, and avoid recording a
heartbeat change.

This PR also makes sure we do not count the first heartbeat record as an event.

Closes #132