This PR applies stricter rule to decide when to trigger a grpc gossip.
It also prevents applying a grpc gossip when the new state
is detected as not being an updated.

It also adds a bunch of defensive asserts.

This changes the condition for the catchup callback to be called.

Now it only happens if the delta provided really triggered a catchup.

blop

Added monotonic invariant. Simplify delta acceptation logic